+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Builtin `!set` now supports int and float options
+
+
 ## [0.6.1] - 2024-10-31
 ### Added
 - Configuration options `call_attempts`, `call_timeout`, and `autoget_timeout`